move the vendored function to utils
